pyproject.toml
[project]
name = "mojo-manual-preprocessing"
version = "0.1.0"
description = "Preprocessing pipeline for Mojo manual documentation to prepare for vector embeddings"
requires-python = ">=3.12"
dependencies = [
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
    "tiktoken>=0.5.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
